requirements.txt
pywinctl
pywinauto
keyboard
pystray
pillow
pygetwindow
requests